Refactor S3GetKeyOperator into separate file
